FF time passes float. Change csv_row type to double

+++ ca.cpp
@@ -136,7 +136,7 @@ int cli_main(int argc, char *argv[]){
     float eff_ff = sp_ff / (num_row_threads * num_col_threads);
 
 
-    float csv_row[] = {board_w, board_h, num_row_threads, num_col_threads, num_iters, \
+    double csv_row[] = {board_w, board_h, num_row_threads, num_col_threads, num_iters, \
                         seq_time, par_time, ff_par_time, sp_par, sp_ff, sc_par, sc_ff, eff_par, eff_ff};
     int row_len = sizeof(csv_row)/sizeof(csv_row[0]);
     char delim[] = ", ";
